Rename zip validation for clarity
